💄 initial dark mode support for settings page

+++ webpack.js
@@ -23,7 +23,8 @@ const fs = require('fs');
 const {APP_EVENTS, ELECTRONIM_VERSION} = require('./src/constants');
 
 const BUNDLES_DIR = 'bundles';
-const ENTRIES = [
+
+const PRELOAD_ENTRIES = [
   'chrome-tabs',
   'help',
   'settings',
@@ -37,9 +38,13 @@ const LIB_ENTRIES = [
   'chrome-tabs/css/chrome-tabs.css',
   'chrome-tabs/css/chrome-tabs-dark-theme.css'
 ];
+const CUSTOM_LIB_ENTRIES = [
+  '/bulma/bulma.dark.scss'
+];
 
-const bundle = webpack({
-  entry: ENTRIES.reduce((acc, entry) => {
+const preloadBundle = webpack({
+  name: 'preload-bundles',
+  entry: PRELOAD_ENTRIES.reduce((acc, entry) => {
     acc[entry] = `/src/${entry}/preload.js`;
     return acc;
   }, {}),
@@ -100,8 +105,9 @@ const bundle = webpack({
   ]
 });
 
-const libBundle = webpack({
-  entry: LIB_ENTRIES.reduce((acc, entry) => {
+const libBundle = ({name, entries}) => webpack({
+  name,
+  entry: entries.reduce((acc, entry) => {
     acc[entry] = entry;
     return acc;
   }, {}),
@@ -110,6 +116,7 @@ const libBundle = webpack({
     assetModuleFilename: '[name][ext]',
     path: path.resolve(__dirname, BUNDLES_DIR, LIB_DIR)
   },
+  mode: 'production',
   target: 'web',
   module: {
     rules: [
@@ -120,6 +127,10 @@ const libBundle = webpack({
       {
         test: /\.(svg|eot|ttf|woff|woff2)$/,
         type: 'asset/resource'
+      },
+      {
+        test: /\.(scss)$/,
+        use: ['style-loader', 'css-loader', 'sass-loader']
       }
     ]
   }
@@ -138,26 +149,32 @@ const toPromise = async webpackBundle => new Promise((resolve, reject) => {
 const exec = async () => {
   console.log('⌛ Starting webpack bundling process...');
   await Promise.all(
-    ENTRIES.map(entry => fsp.access(path.resolve(__dirname, 'src', entry, 'preload.js'), fs.constants.R_OK)),
-    LIB_ENTRIES.map(entry => fsp.access(path.resolve(__dirname, 'node_modules', entry), fs.constants.R_OK))
+    PRELOAD_ENTRIES.map(entry => fsp.access(path.resolve(__dirname, 'src', entry, 'preload.js'), fs.constants.R_OK)),
+    LIB_ENTRIES.map(entry => fsp.access(path.resolve(__dirname, 'node_modules', entry), fs.constants.R_OK)),
+    CUSTOM_LIB_ENTRIES.map(entry => fsp.access(path.resolve(__dirname, entry.substring(1)), fs.constants.R_OK))
   );
   console.log('✅ Required files exist');
-  const bundlesDir = path.resolve(__dirname, BUNDLES_DIR);
-  await Promise.all([bundlesDir].map(dir => fsp.rm(dir, {recursive: true, force: true})));
-  console.log('🧹 Cleaned previous build...');
-  const bundleStats = await toPromise(bundle);
-  console.log(`⏱️ Webpack bundling completed in ${bundleStats.endTime - bundleStats.startTime}ms`);
-  const libBundleStats = await toPromise(libBundle);
-  console.log(`⏱️ Webpack lib bundling completed in ${libBundleStats.endTime - libBundleStats.startTime}ms`);
-  if (bundleStats.hasErrors()) {
-    console.error('⚠️ Bundling errors found:');
-    bundleStats.compilation.errors.forEach(error => console.error(error));
-  } else if (libBundleStats.hasErrors()) {
-    console.error('⚠️ Bundling errors found for lib:');
-    libBundleStats.compilation.errors.forEach(error => console.error(error));
-  } else {
-    console.log('🚀 Bundling completed successfully');
-    console.log(bundleStats.toString({colors: true}));
+  const bundles = [preloadBundle];
+  if (!process.argv.includes('--no-lib')) {
+    const bundlesDir = path.resolve(__dirname, BUNDLES_DIR);
+    await Promise.all([bundlesDir].map(dir => fsp.rm(dir, {recursive: true, force: true})));
+    console.log('🧹 Cleaned previous build...');
+    bundles.push(libBundle({name: 'lib', entries: LIB_ENTRIES}));
+    bundles.push(libBundle({name: 'custom-lib', entries: CUSTOM_LIB_ENTRIES}));
+  }
+  let hasErrors = false;
+  for (const bundlePromise of bundles.map(toPromise)) {
+    const stats = await bundlePromise;
+    console.log(`⏱️ Webpack ${stats.compilation.name} bundling completed in ${stats.endTime - stats.startTime}ms`);
+    if (stats.hasErrors()) {
+      hasErrors = true;
+      console.error(`⚠️ Bundling errors found for ${stats.compilation.name}:`);
+      stats.compilation.errors.forEach(error => console.error(error));
+    }
+    if (!hasErrors) {
+      console.log('🚀 Bundling completed successfully');
+      console.log(stats.toString({colors: true}));
+    }
   }
 };
